requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorator==5.2.1
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.45
# google-ai-generativelanguage==0.6.18
# google-api-core==2.25.1
# google-api-python-client==2.178.0
# google-auth==2.40.3
# google-auth-httplib2==0.2.0
# google-generativeai==0.8.5
# googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langsmith==0.4.13
MarkupSafe==3.0.2
moviepy==2.2.1
narwhals==2.0.1
numpy==2.2.6
opencv-python==4.12.0.88
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.5
watchdog==6.0.0
webcolors==24.11.1
zstandard==0.23.0
